Jörg Sommer: 25819: improve module parameter completion.
First, do not embedd the = in the parameter by define it a
parameter‐value separator; -S =.

Second, look for already given parameters in all words; -w.

Third, use the curcontext variable; -C.

Forth, distinct between boolean and non‐boolean parameters, i.e. do not
append a = to the end. And include the type of the parameter in the help
message. This information is only given in the full output of modinfo,
not in the narrowed output with --field.