31906: fix race-condition interaction of $pipestatus with job control
printjob() should not reference oldjobtab for job numbers unless it is being called from bin_fg().  printjob() also must not attempt to update pipestats when called from bin_fg().  acquire_pgrp() should not loop infintely if the shell is not interactive.  Update the $pipestatus stress test so that it also exercises the oldjobtab repair.