38153: change the way long options are examined for "=" signs to more accurately identify options that take arguments
Note in workers/39135:
The pattern in 266/268 appears to match the '=' sign even though it's in the
second colon-separated field, where it doesn't denote a mandatory argument.
I'm guessing the pattern match on lines 266/268 should be fixed ...

This commit alters those patterns in what is hoped to be the desired way.