39847: Update helper function name pattern
In 580698bc71, Oliver updated a whole bunch of options in a number of
completions. While at it, he also changed a bit of coding style, to
better adhere to Etc/zsh-development-guide — among those changes is to
convert "function foobar() {" function definitions to "foobar() {".

This broke Util/check-tmux-state so that it couldn't detect any
supported tmux sub-commands within _tmux anymore. This updates the
regular expression that is used in the script to match and list these
function definitions.