-Dproperty=/path_complete
I think this change is desirable since line 86 already proposes a _default completion and path completion if magic_equal_subst is enabled and it is never reached if -D is not matched on -D* e.g. "-Dproperty" as most people use -D properties as I'm aware (Not a Java programmer).