Specify system python in dnf completion
The python executable in the python_exec variable was the python[2|3]
executable found in $PATH. When the python[2|3] executable was not the
system installed version the completion command raised an `ImportError`
as the 'dnf' module was not found.

This commit just completely specifies the python version to be the
system version `/usr/bin/python[2|3]`, allowing completion within a
python dev environment.