42324: _git: handle mutually exclusive options
Currently `git show --stat -p` is not considered a valid completion and
prevents further completions like `--no-abbrev`.

It is a valid use case however to use `--stat` with `-p` (and there is a
synonym `--patch-with-stat` for it even).

This patch changes `__git_setup_diff_options` to only handle mutually
exclusive arguments reported by Git (as a fatal error) in a special way.