44722: _make: use --always-make instead of .PHONY for GNU make
Using ".PHONY" might actually execute (phony) targets then, while
"--always-make" triggers processing of all targets, respecting `-n`
(`--dry-run`).
(This can be seen with `make -nsp .PHONY` in
https://github.com/neomake/neomake/tree/76b110d49a, where it actually
causes infinite processing then)

".PHONY" was used since the beginning here (37012f06a [1]).

As for bmake (which also uses `.PHONY`): there does not appear to be a
similar option for it.
While it might work better there (i.e. not causing targets to be run),
it causes an error at least when testing it quickly - which indicates
that it behaves as if using any other non-existing (or special) target
there maybe:

> bmake: don't know how to make .PHONY. Stop

This was added in ecc0a5ece [2] (workers-39654).

Since it uses `-s -f "$file" -V.ALLTARGETS` with `bmake` before to get
targets explicitly, there might not be a reason for calling `bmake` with
`-nsdg1Fstdout … .PHONY` additionally in the first place?!

1: https://github.com/danielshahaf/zsh/commit/37012f06a
2: https://github.com/danielshahaf/zsh/commit/ecc0a5ece