Partial fix for handling of tied arrays.
As of this commit when a tied array is declared but neither the scalar
nor the array has an initializer, the array is initialized to empty.

The scalar struct param of a tied pair stores a direct pointer to the
internal array value of the array struct param, and upon assignment
modifies it without referring to the containing struct.  This means
that there's no opportunity to clear the PM_DECLAREDNULL bits on both
structs when the scalar is assigned.  Conversely, assigning to the
array does use the struct for the scalar.