48131/0003: FAQ: 3.31: Move two paragraphs unchanged, reordering them
Factored out from the next patch for readability of the diff.

The answer, as it stands in this commit, doesn't flow correctly.  In
order to keep the diffs small and easy to follow, this commit simply
adds a TODO to record this issue, a TODO which is fixed by the next
commit in this series (the child of this commit).  Thus, a request:

DEAR PEOPLE FROM THE FUTURE: Please don't create new branches off this
commit; create them off the parent commit or off the child commit
instead.  Thank you.