github #119: fix completion for git bisect visualize
When trying to use the completion for git bisect visualize one currently
is greeted with the curious error of something like this:

    git bisect visualize _git-bisect:89: command not found: arch

While in this case 'arch' is the name of the first folder in the
directory. The cause for this is a missing separation of switch cases
which we fix by inserting the needed ";;".

Signed-off-by: Christian Heusel <christian@heusel.eu>