Completion: Add make variables with values
So far only variable names were added. However Makefiles and make would
also return possible values which are not completed so far. The values
are stored and overwritten in the map without using them anywhere.
So instead it currently makes more sense to complete the whole variable
name-value combination instead of only the name.