github #129: fix _gpg completion
1) As for now, `--local-user` completion is irrelevant: it completes system users, while gpg expects private keys IDs

2) `secret-keys` completion was also broken and completed nothing.
The reason of that to happen was in that fact that it assumed `uid` token would be next after `fpr` (same as it do in public-keys completion), while in current gnupg versions there is at least `grp` token, and potentially can be others.
So, instead i+=2 I made `until "uid"` loop.