print selection list with continue --print-selections
The select command only reprints the selection list when the input is
empty. There was no way to explicitly reprint the list in response to
a specific user input, such as:

    1) <m>ouse  2) <c>ow  3) <t>igger 4) <?>SHOW THIS MENU AGAIN
    ?#

This change extends the continue builtin with a new option:
--print-selections. When this option is used, the select command
reprints the menu before showing the PS3 prompt on the next iteration.

    case "$REPLY" in
    ...
        ('?') continue --print-selections;;
    ...
    esac

Signed-off-by: Masatake YAMATO <yamato@redhat.com>